Moved select command into GameController as the board shouldn't be keeping track of this information
